Trait embedded_graphics_framebuf::backends::FrameBufferBackend

pub trait FrameBufferBackend {
    type Color: PixelColor;

    fn set(&mut self, index: usize, color: Self::Color);
    fn get(&self, index: usize) -> Self::Color;
    fn nr_elements(&self) -> usize;
}

This trait marks the requirements for backends for a FrameBuf.

In a basic scenario this is just some memory. But one could implement more elaborate backends which allow manipulation of the data on the fly.

Required Associated Types

type Color: PixelColor

Required Methods

fn set(&mut self, index: usize, color: Self::Color)

Sets a pixel to the respective color

fn get(&self, index: usize) -> Self::Color

Returns a pixels color

fn nr_elements(&self) -> usize

Nr of elements in the backend