Struct embedded_graphics_framebuf::FrameBuf

pub struct FrameBuf<C: PixelColor, B: FrameBufferBackend<Color = C>> {
    pub data: B,
    /* private fields */
}

Constructs a frame buffer in memory. Lets you define the width(X), height (Y) and pixel type your using in your display (RGB, Monochrome etc.)

Example

use embedded_graphics::{
    mono_font::{ascii::FONT_10X20, MonoTextStyle},
    pixelcolor::Rgb565,
    prelude::*,
    text::Text,
};
use embedded_graphics_framebuf::FrameBuf;

// Create a framebuffer for a 16-Bit 240x135px display
let mut data = [Rgb565::BLACK; 240 * 135];
let mut fbuff = FrameBuf::new(&mut data, 240, 135);

// write "Good luck" into the framebuffer.
Text::new(
    &"Good luck!",
    Point::new(10, 13),
    MonoTextStyle::new(&FONT_10X20, Rgb565::WHITE.into()),
)
.draw(&mut fbuff)
.unwrap();

Fields
data: B
Implementations

impl<C: PixelColor, B: FrameBufferBackend<Color = C>> FrameBuf<C, B>

pub fn new(data: B, width: usize, height: usize) -> Self

Create a new FrameBuf on top of an existing memory slice.

Panic

Panics if the size of the memory does not match the given width and height.

Example

use embedded_graphics::{pixelcolor::Rgb565, prelude::RgbColor};
use embedded_graphics_framebuf::FrameBuf;
let mut data = [Rgb565::BLACK; 240 * 135];
let mut fbuff = FrameBuf::new(&mut data, 240, 135);

pub fn width(&self) -> usize

Get the framebuffers width.

pub fn height(&self) -> usize

Get the framebuffers height.

pub fn set_color_at(&mut self, p: Point, color: C)

Set a pixel’s color.

pub fn get_color_at(&self, p: Point) -> C

Get a pixel’s color.

impl<C: PixelColor + Default, B: FrameBufferBackend<Color = C>> FrameBuf<C, B>

pub fn reset(&mut self)